+++ src/models/properties.ts
@@ -624,6 +624,10 @@ export interface StorageMeta {
      */
     storeUrl?: boolean,
 
+    /**
+     * Post process the path
+     */
+    postProcess?: (pathOrUrl: string) => Promise<string>
 }
 
 /**